requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client for Supabase integration
httpx==0.25.2

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Integration with existing services
# Note: These should match the existing project requirements
# crewai>=0.1.0  # Uncomment if needed
# langchain>=0.1.0  # Uncomment if needed
